+++ app/src/main/java/com/teampophory/pophory/domain/repository/photo/PhotoRepository.kt
@@ -6,14 +6,14 @@ import com.teampophory.pophory.data.network.model.album.PhotoListResponse
 import com.teampophory.pophory.domain.model.S3Image
 
 interface PhotoRepository {
-    suspend fun getPhotos(id: Int): Result<PhotoListResponse>
+    suspend fun getPhotos(id: Long): Result<PhotoListResponse>
 
     suspend fun deletePhoto(photoId: Long): Result<Unit>
 
     suspend fun getStudios(): Result<List<Studio>>
 
     suspend fun addPhotoToPophory(
-        albumId: Int,
+        albumId: Long,
         takenAt: String,
         studioId: Long,
         fileName: String,
